backend-drf/requirements.txt
-i https://pypi.org/simple
asgiref==3.11.0; python_version >= '3.9'
django==6.0.1; python_version >= '3.12'
django-cors-headers==4.9.0; python_version >= '3.9'
djangorestframework==3.16.1; python_version >= '3.9'
djangorestframework-simplejwt==5.5.1; python_version >= '3.9'
pyjwt==2.10.1; python_version >= '3.9'
sqlparse==0.5.5; python_version >= '3.8'
tzdata==2025.3; python_version >= '2'
